+++ src/lib.rs
@@ -45,7 +45,6 @@ pub mod parser;
 pub mod poll;
 pub mod post;
 pub mod profile;
-pub mod reply;
 pub mod threading;
 pub mod tokenizer;
 pub mod util;